Fix mistake with project field in new todo field
This commit fixes a small mistake where, when entering a project name
that does not match an existing project name, a dialog will pop up
saying that the todo item will be placed in "Default" but then will be
placed in a new project matching the name entered into the project
field of the new todo form. The dialog box has been changed to indicate
the creation of a new project.